fix: use long in shop settings address

+++ features/merchant/shop_common/src/main/java/com/tokopedia/shop/common/graphql/data/shoplocation/ShopLocationModel.kt
@@ -19,19 +19,19 @@ data class ShopLocationModel (
     val address: String = "",
     @SerializedName("districtId")
     @Expose
-    val districtId: Int = -1,
+    val districtId: Long = -1,
     @SerializedName("districtName")
     @Expose
     val districtName: String = "",
     @SerializedName("cityId")
     @Expose
-    val cityId: Int = -1,
+    val cityId: Long = -1,
     @SerializedName("cityName")
     @Expose
     val cityName: String = "",
     @SerializedName("stateId")
     @Expose
-    val stateId: Int = -1,
+    val stateId: Long = -1,
     @SerializedName("stateName")
     @Expose
     val stateName: String = "",
